pyproject.toml
[project]
name = "c5-image-caption"
version = "0.1.0"
description = "Modular image captioning framework on VizWiz with swappable encoders, decoders, and text representations."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "evaluate>=0.4.0",
    "huggingface-hub>=1.7.1",
    "joblib>=1.3.0",
    "kaleido>=0.2.1",
    "matplotlib>=3.10.8",
    "natsort>=8.4.0",
    "nltk>=3.8",
    "numpy>=2.4.3",
    "opencv-python>=4.13.0.92",
    "optuna>=4.0.0",
    "pillow>=12.1.1",
    "plotly>=5.18.0",
    "pyyaml>=6.0.3",
    "rouge-score>=0.1.2",
    "scikit-learn>=1.8.0",
    "seaborn>=0.13.2",
    "tokenizers>=0.21.0",
    "torchsummary>=1.5.1",
    "torchviz>=0.0.3",
    "tqdm>=4.67.3",
    "transformers>=5.3.0",
    "wandb>=0.19.0",
    "peft>=0.15.0",
    "accelerate>=1.0.0",
    "bitsandbytes>=0.45.0",
    "qwen-vl-utils>=0.0.8",
    "flash-linear-attention>=0.4.2",
    "torch==2.8.0",
    "torchvision==0.23.0",
    "torchaudio==2.8.0",
]

[project.scripts]
c5-caption = "main:main"

[[tool.uv.index]]
name = "pytorch-cu129"
url = "https://download.pytorch.org/whl/cu129"
explicit = true

[tool.uv.sources]
torch = [
  { index = "pytorch-cu129", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
torchvision = [
  { index = "pytorch-cu129", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
